Fix handling of symbols in kernel modules.
Since we are using the symbol table from the backing ELF files and
not the in-kernel symbol table (as is done on illumos), we need to
relocate symbol addresses.  Note that we can't simply add the base
address of the module to the symbol values as the kernel linker
"packs" the sections in.  Instead, find the lowest offset and subtract
that from the base address to determine the relative address for each
module.  Technically this isn't quite right but works well enough in
practice.

Note that I have to use -M to force-load module symbols currently.
This may be due to me not understanding the mdb syntax well enough to
force on-demand symbol loading.