Handle symbols for global labels in assembly.
In particular, the symbol table entry for labels defined as @function
without an end in FreeBSD end up with a size of zero.  However, we want
to find these labels (and depend on them to identify trapframes).  To
that end, synthesize a size for these labels by claiming addresses from
the symbol's value up to the next symbol in the symbol table.