Handle tail calls in the '%a' format used for printing stack traces.
Functions that use tail calls may perform the tail call as the last
instruction of the function.  In this case, the return PC saved on
the stack will reference the first instruction after the end of the
function.  As a result, a symbol lookup on the PC will fail to find a
valid symbol.  Work around this by examining 'addr - 1' if the symbol
lookup for 'addr' fails.  If 'addr - 1' has a valid symbol which is a
function, use that symbol for the address.

This fixes the output of tail calls in the 'panic' and 'vpanic' functions
in panic stack traces on FreeBSD/amd64.