Properly handle endianness detection under FreeBSD.
FreeBSD always defines _BIG_ENDIAN and _LITTLE_ENDIAN and then defines
a machine-specific _BYTE_ORDER which is meant to be compared against
_BIG_ENDIAN or _LITTLE_ENDIAN.  illumos only defines _BIG_ENDIAN on
BE platforms.  As a result, code using '#ifdef _BIG_ENDIAN' was always
enabled on FreeBSD.  To cope, #undef _BIG_ENDIAN on FreeBSD non-BE
platforms at the start of files.  This is less invasive than patching
each #ifdef to compare against _BYTE_ORDER.

Among other things, this fixes decoding of bitfield values in ::print.